Build Properly from source
Rather than just pulling down whatever has been built on the alpine
edge repo, instead, using a git submodule at the latest tag, build from
source as part of a staged docker build.

Remove all the default env var values from the Dockerfile, and instead
require env vars to be passed when executing (to ensure that all config)
is deliberate. To remove some pain, though, I've given some defaults for
the ports.

Modify the entry script to exec into the binary instead of just running
it. This ensures that it's running as PID 1, and that the sslh receives
any signals (ie. when running `docker stop`).

Add in some labels to the image, for transparency's sake.

Finally, run sslh as its own user, not root. Because best practices.